Fixes expression generation mistakes and c2llvm.py
Expression generation used the lhs for the rhs (copy-paste mistake) as
well as using the wrong index to get the previous temporary registers.

c2llvm.py had been mistakenly changed back to previous state, now the
generated files are placed in the correct directory and some simple
error messages are printed to the console.